user/security: replace DaoAuthenticationProvider with a custom provider
The DaoAuthenticationProvider (DAP) is inadequate for our use-case, as
can be seen by the hack that was necessary to get it to work.

It is an authentication provider that assumes
UsernamePasswordAuthenticationToken, a type of token that contains only
a username and password, but we have an extra 'type' field.  Worse yet,
it constraints the username to be a String, despite it not being so in
the token itself.  This would all be tolerable, but
AbstractUserDetailsAuthenticationProvider, the superclass of DAP goes a
step further to ruin any possibility of it being useful in our case by
forcing 'username' into a string before passing it to DAP#retrieveUser
which, in turn, passes it to the UserDetailsService we used to
implement.

As if out of spite, the CustomUserDetailsService DAO does not get
properly informed of what to handle.

In the end, best we can do is scratch DAP and use
AbstractUserDetailsAuthenticationProvider directly, but even then, we'd
need a non-insignificant amount of hackery, because we'd still be
dealing with its 'username' deusefulnesification and hard-coding of
UsernamePasswordAuthenticationToken.

Scratch that, just implementing our own AuthenticationProvider is
easier.

Closes: https://github.com/RAF-SI-2024/Banka-4-Backend/issues/30